fx-youtube-insights/requirements.txt
# Core dependencies
google-api-python-client==2.97.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
pandas==2.1.1
duckdb==0.9.2
python-dotenv==1.0.0
streamlit==1.28.0
pydantic==2.4.2
tenacity==8.2.3
pytz==2023.3.post1
requests==2.31.0
scipy==1.11.3
plotly==5.17.0
PyYAML==6.0.1